SOLR-13101: Address flakiness of tests using async pulls and handle interrupt properly
