+++ solr/core/src/test/org/apache/solr/store/shared/SharedStoreMissingCoreTest.java
@@ -235,26 +235,4 @@ public class SharedStoreMissingCoreTest extends SolrCloudSharedStoreTestCase {
     Collections.sort(actualDocs);
     assertEquals("wrong docs", expectedDocs.toString(), actualDocs.toString());
   }
-
-  private Map<String, CountDownLatch> configureTestBlobProcessForNode(JettySolrRunner runner) {
-    Map<String, CountDownLatch> asyncPullTracker = new HashMap<>();
-
-    CorePullerFeeder cpf = new CorePullerFeeder(runner.getCoreContainer()) {
-      @Override
-      protected CorePullTask.PullCoreCallback getCorePullTaskCallback() {
-        return new PullCoreCallback() {
-          @Override
-          public void finishedPull(CorePullTask pullTask, BlobCoreMetadata blobMetadata, CoreSyncStatus status,
-                                   String message) throws InterruptedException {
-            CountDownLatch latch = asyncPullTracker.get(pullTask.getPullCoreInfo().getCoreName());
-            latch.countDown();
-          }
-        };
-      }
-    };
-
-    BlobProcessUtil testUtil = new BlobProcessUtil(runner.getCoreContainer(), cpf);
-    setupTestBlobProcessUtilForNode(testUtil, runner);
-    return asyncPullTracker;
-  }
 }
